+++ wikitextparser/_template.py
@@ -313,6 +313,12 @@ class Template(SubWikiTextWithArgs):
                 return True
         return False
 
+    def del_arg(self, name: str) -> None:
+        """Delete all arguments with the given then."""
+        for arg in reversed(self.arguments):
+            if arg.name.strip(WS) == name.strip(WS):
+                del arg[:]
+
 
 def mode(list_: List[T]) -> T:
     """Return the most common item in the list.