+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 master branch
 -------------
+- Add ``del_arg`` method.
 - Improved the ``lists`` method of ``Template`` and ``Argument`` classes.
 - Fixed a bug in detection of multiline arguments. (#13)
 - Deprecated ``capital_links`` parameter of ``Template.normal_name``. Use